requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
streamlit
pandas
plotly
requests